refactor(forcing): invert the from_bundles arrow into a forcing-side engine
Move the forcing-assembly science out of jcm.runners into a new
jcm.forcing_assembly module next to the readers it drives: the auto-emission
resolution, the ozone/emissions/dms/dust/oxidant/MACv2 attach chain, the
per-product {year}/hf:// resolution and the merge-compatibility guard. Both
doors now go through one engine — ForcingData.from_bundles drives
forcing_assembly.build_forcing directly (arrow inverted; it no longer depends
on the runner's build), and runners.build_forcing is a thin delegate to it. The
pySES column branch (attach_jam_forcing) stays in the runner, deliberately not
unified, and shares the resolution helpers. The two runner utilities the
equivalence test patches — _resolve_data_path and _resolve_auto_ozone — stay in
the runner and are called back through it, so a mock.patch reaches both doors;
the moved names are re-exported from runners so existing runners.<name> call
sites/patch targets keep resolving.

Behaviour-preserving: the #763 forcing/runners suites pass unchanged. The one
test-mechanics change is the from_bundles capture helper, retargeted from
runners.build_forcing to forcing_assembly.build_forcing (the new pre-resolution
interception point) — its assertions are unchanged.
