+++ jcm/runners.py
@@ -30,8 +30,6 @@ from omegaconf import DictConfig
 
 from jcm import provenance
 from jcm.data import bundle_names
-from jcm.data import input_resolution as ir
-from jcm.data import mirror_manifest as mm
 from jcm.diffusion import DiffusionFilter
 from jcm.forcing import expand_yearly_files
 from jcm.initial_states import (
@@ -512,47 +510,6 @@ def _resolve_data_path(path):
 _expand_years = expand_yearly_files
 
 
-def _forcing_products(file_spec, years, available):
-    """Split a forcing spec into per-product, year-expanded file sets.
-
-    Thin adapter over :func:`jcm.data.input_resolution.forcing_products` — the
-    single home for the list-into-products split + ``{year}`` coverage clamp.
-    """
-    return ir.forcing_products(file_spec, years, available)
-
-
-def _open_forcing_dataset(path):
-    """Open one product's file(s) as a single xarray dataset.
-
-    A product may be a list of yearly files (a ``{year}`` expansion) — those
-    share one time axis and are concatenated with
-    ``open_mfdataset(combine="by_coords")``. A scalar path opens directly.
-    """
-    import xarray as xr
-    if isinstance(path, (list, tuple)):
-        paths = [str(p) for p in path]
-        return (xr.open_mfdataset(paths, combine="by_coords") if len(paths) > 1
-                else xr.open_dataset(paths[0]))
-    return xr.open_dataset(str(path))
-
-
-def _product_available_years(forcing_cfg, key: str):
-    """Per-product source coverage, falling back to ``available_years``.
-
-    A preset can mix yearly products with different coverages (e.g.
-    ``forcing_era5`` surface files run to 2024 while the FZJ ``ozone_amip``
-    and CEDS ``emissions_amip`` bundles both end in 2022); a per-product
-    override (``ozone_available_years`` / ``emissions_available_years`` /
-    ``oxidants_available_years``) keeps each pattern's expansion inside the
-    files that actually exist — for run dates beyond it, the time lookup
-    clamps to the last sample. ``key`` selects the override; any product
-    without one shares the top-level ``available_years``.
-    """
-    avail = forcing_cfg.get(key, None)
-    return avail if avail is not None else forcing_cfg.get(
-        "available_years", None)
-
-
 def build_terrain(cfg: DictConfig, coords) -> TerrainData:
     terrain_cfg = cfg.terrain
     kind = terrain_cfg.kind
@@ -920,311 +877,166 @@ def _pyses_lid_sponge_term(dycore, sponge_cfg):
 # Forcing
 # ---------------------------------------------------------------------------
 
-#: The ``auto`` emission-bundle naming convention lives in
-#: :mod:`jcm.data.bundle_names` (a jcm-import-free module) so the build-time
-#: resolver here and the benchmark's pre-GPU prefetch enumerator share one
-#: source of truth. ``auto`` resolves to the per-grid HF bundle when a
-#: prognostic-aerosol (JAM) package is active.
-_EMISSION_AUTO_BUNDLES = bundle_names.EMISSION_AUTO_BUNDLES
-
-
-def _emission_auto_resolves_to_none(key, coords, jam, is_pyses) -> bool:
-    """Whether ``auto`` for one emission ``key`` resolves to None — pure classify.
-
-    ``auto`` yields None (no bundle) on the pySES path, for a non-JAM package
-    (neither consumes prescribed emissions), or when the mirror does not publish
-    THIS key's bundle for the model grid — a check that is now key-specific
-    (:func:`jcm.data.bundle_names.bundle_is_published`): the horizontal token
-    must be a published grid, and the level-dependent ``oxidants_file`` bundle
-    additionally requires a published layer count AND a hybrid vertical (the
-    bundle is on hybrid-level pressures — mapping it level-for-level onto a
-    sigma grid is silently wrong), so a published-horizontal / unpublished-level
-    combo (e.g. ``t63_l8``) OR a sigma grid that merely shares a published
-    (token, nlev) (e.g. ``echam_t42_l8_sigma`` at ``t63``/``l47``) nulls
-    ``oxidants_file`` while the level-free emissions/dms/dust keys still resolve
-    (F2). No fetch, no side
-    effects. Shared by the build-time resolver
-    (:func:`_resolve_one_emission_input`) and the config-trap warner
-    (:func:`warn_on_config_traps`) so the None-decision the warning reasons
-    about can never drift from the one the build actually applies. The pySES /
-    non-JAM cases return before ``_grid_token`` so pySES native coords are never
-    asked for a spectral token.
-    """
-    if is_pyses or not jam:
-        return True
-    # Publication gating is the manifest lookup ``mm.is_published`` (the whole
-    # product table), which mirror_manifest generalises from — and is
-    # cross-checked against — ``bundle_names.bundle_is_published`` (#751).
-    manifest = mm.load_manifest()
-    return not mm.is_published(
-        manifest, mm.product_for_key(manifest, key),
-        _grid_token(coords), int(coords.nodal_shape[0]), _vertical_kind(coords))
-
-
-def _resolve_one_emission_input(value, key, coords, jam, is_pyses):
-    """Resolve one prescribed-emission forcing value (``auto`` / explicit).
-
-    * ``auto`` → the per-grid HF bundle, resolved + eager-fetched *now* (cold
-      cache fails loudly at build time) by :func:`jcm.data.input_resolution.
-      resolve_input` — but only on the spectral path with a JAM package active;
-      pySES native grids are not the spectral-token bundles and a non-JAM
-      package consumes no emissions, so both give ``None`` (the early gate keeps
-      pySES native coords from being asked for a spectral token). Fetch is
-      routed through ``_resolve_data_path`` so the hf:// scheme + provenance are
-      honoured. ``auto`` is the only grid-portable mechanism; there is no
-      user-facing ``{grid}``/``{nlev}`` template.
-    * explicit null (``None``/``""``/``"null"``) → ``None`` (opt-out).
-    * an explicit path / ``hf://`` URL → returned unchanged for the attach
-      helper's own ``_resolve_data_path`` to fetch. A ``{year}`` pattern is
-      left intact for :func:`_expand_years` (there is no ``{grid}``/``{nlev}``
-      substitution — use ``auto`` to let a config follow the grid).
-    """
-    if value == "auto":
-        if _emission_auto_resolves_to_none(key, coords, jam, is_pyses):
-            # Non-mirrored grid / unpublished level / pySES / non-JAM: resolve
-            # ``auto`` to None so the run falls back to the emission-free
-            # baseline instead of aborting on a 404. ``warn_on_config_traps``
-            # surfaces the silent degrade (same None-decision), not a log here.
-            return None
-        r = ir.resolve_input(
-            key, "auto", grid_token=_grid_token(coords),
-            nlev=int(coords.nodal_shape[0]), vertical=_vertical_kind(coords),
-            fetch=lambda rel: _resolve_data_path("hf://" + rel))
-        return r.paths[0] if not r.is_none else None
-    if value in (None, "", "null", "none"):
-        return None
-    return value
-
-
-def _resolve_emission_inputs(forcing_cfg, cfg, coords, is_pyses):
-    """Concretise the ``auto`` emission keys on ``forcing_cfg``.
-
-    Returns ``forcing_cfg`` with the four prescribed-emission keys resolved (see
-    :func:`_resolve_one_emission_input`). ``auto`` is keyed off whether the
-    composed physics is the prognostic-aerosol (JAM) module — the only package
-    that consumes prescribed emissions.
-
-    Pure resolution only: the silent degrade when a JAM run's ``auto`` keys
-    null on a non-mirrored grid (or the pySES path) is surfaced by
-    :func:`warn_on_config_traps`, which folds it into ONE config-trap warning
-    with the zero-emission finding rather than an invisible info log here (F2).
-    """
-    from omegaconf import OmegaConf
-
-    jam = str((cfg.get("physics", {}) or {}).get("aerosol_module", "")) == "jam"
-    updates = {
-        key: _resolve_one_emission_input(
-            forcing_cfg.get(key, None), key, coords, jam, is_pyses)
-        for key in _EMISSION_AUTO_BUNDLES
-    }
-    return OmegaConf.merge(forcing_cfg, updates)
+# The forcing-assembly science — ``auto`` resolution, the attach chain, the
+# merge-compatibility guard — lives in :mod:`jcm.forcing_assembly` next to the
+# readers it drives; the runner keeps only the cfg dispatch (``build_forcing``)
+# plus the pySES column-sampling branch. These names are re-exported so existing
+# ``runners.<name>`` call sites and test patch targets keep resolving.
+from jcm.forcing_assembly import (  # noqa: E402
+    _assert_uniform_time_axis as _assert_uniform_time_axis,
+    _attach_dms as _attach_dms,
+    _attach_dust as _attach_dust,
+    _attach_emissions as _attach_emissions,
+    _attach_macv2_weights as _attach_macv2_weights,
+    _attach_oxidants as _attach_oxidants,
+    _attach_ozone as _attach_ozone,
+    _emission_auto_resolves_to_none as _emission_auto_resolves_to_none,
+    _ensure_parent_forcing as _ensure_parent_forcing,
+    _forcing_products as _forcing_products,
+    _grid_token as _grid_token,
+    _merge_disjoint_emissions as _merge_disjoint_emissions,
+    _model_latlon_deg as _model_latlon_deg,
+    _open_forcing_dataset as _open_forcing_dataset,
+    _product_available_years as _product_available_years,
+    _product_time_axis as _product_time_axis,
+    _reject_year_pattern as _reject_year_pattern,
+    _resolve_emission_inputs as _resolve_emission_inputs,
+    _resolve_one_emission_input as _resolve_one_emission_input,
+    _resolve_oxidant_paths as _resolve_oxidant_paths,
+    _resolve_pyses_emission_paths as _resolve_pyses_emission_paths,
+    _vertical_kind as _vertical_kind,
+    assemble_spectral_forcing as assemble_spectral_forcing,
+)
 
 
 def build_forcing(cfg: DictConfig, coords, dycore=None):
-    """Build a ``ForcingData`` from ``cfg.forcing``.
+    """Build a ``ForcingData`` from ``cfg.forcing`` (thin CLI-door delegate).
 
-    ``kind: default`` returns ``None`` — ``Model.run`` then falls back to the
-    aquaplanet ``default_forcing(coords.horizontal)``. ``kind: from_file``
-    loads a netCDF boundary file via ``ForcingData.from_file``.
-
-    Optionally attaches an ozone climatology (``cfg.forcing.ozone_file``),
-    prescribed aerosol emissions (``cfg.forcing.emissions_file``), a seawater
-    DMS climatology (``cfg.forcing.dms_file``), a dust source/erodibility map
-    (``cfg.forcing.dust_file``) and an oxidant climatology
-    (``cfg.forcing.oxidants_file``); all files must already be on the model
-    horizontal grid (the HAMMOZ-style natural-emission files may have
-    descending latitude — they are validated and flipped to model order).
-
-    The four prescribed-emission keys default to ``auto`` (see
-    ``forcing/default.yaml``): a JAM (online-aerosol) package then composes the
-    per-grid HF emission bundles by itself, so
-    ``physics=echam-jam grid=echam_t63_l47_hybrid`` is the documented-canonical
-    run without nine lines of ``hf://`` overrides. Non-JAM packages and the
-    pySES backend leave them empty; ``forcing.<key>=null`` opts out explicitly.
+    The engine lives forcing-side in :func:`jcm.forcing_assembly.build_forcing`
+    (the same one the Python door ``ForcingData.from_bundles`` drives, so the two
+    provably agree, #751); this door only forwards the composed ``cfg``. The
+    pySES column branch it dispatches to (:func:`_build_pyses_forcing`) stays in
+    the runner — deliberately not unified with the spectral assembly.
     """
-    is_pyses = dycore is not None and hasattr(dycore, "colmap")
-    _forcing_cfg = cfg.get("forcing", None)
-    if _forcing_cfg is not None:
-        _forcing_cfg = _resolve_emission_inputs(
-            _forcing_cfg, cfg, coords, is_pyses)
+    from jcm import forcing_assembly as fa
+    return fa.build_forcing(cfg, coords, dycore=dycore)
 
-    if is_pyses:
-        # pySES backend: monthly lon/lat climatology + JAM aerosol inputs,
-        # each bilinearly interpolated onto the physics columns at build
-        # time by ``jcm.dycore.pyses.forcing`` (files may live on any
-        # regular lon/lat grid). ``ozone_file: auto`` resolves the packaged
-        # climatology — column sampling has no exact-grid requirement, so
-        # the T63 file serves any pySES resolution.
-        from jcm.dycore.pyses.forcing import build_forcing as pyses_build_forcing
 
-        ozone_file = _forcing_cfg.get("ozone_file", None)
-        if ozone_file == "auto":
-            from importlib import resources
+def _build_pyses_forcing(_forcing_cfg, dycore, coords):
+    """Build pySES-backend forcing: bilinear column sampling of the inputs.
 
-            cand = (Path(str(resources.files("jcm")))
-                    / "data" / "bc" / "t63" / "ozone.nc")
-            if cand.exists():
-                ozone_file = str(cand)
-            else:
-                logging.warning(
-                    "forcing.ozone_file=auto: packaged t63/ozone.nc missing "
-                    "— pySES run falls back to the ANALYTIC ozone profile "
-                    "(~12 W/m2 clear-sky OLR low bias)."
-                )
-                ozone_file = None
-        elif ozone_file in ("", "null", "none"):
+    Kept in the runner (not unified with the spectral assembly, #751): the pySES
+    core interpolates every gridded field (ozone / emissions / dms / dust /
+    oxidants) onto its physics columns at build time via ``attach_jam_forcing``
+    (files may live on any regular lon/lat grid — no exact-grid requirement), so
+    it delegates to ``jcm.dycore.pyses.forcing.build_forcing`` rather than the
+    dinosaur attach helpers. It DOES share the resolution helpers
+    (``_resolve_pyses_emission_paths`` / ``_resolve_oxidant_paths`` / year
+    expansion) so the two paths cannot drift.
+    """
+    from jcm.dycore.pyses.forcing import build_forcing as pyses_build_forcing
+
+    ozone_file = _forcing_cfg.get("ozone_file", None)
+    if ozone_file == "auto":
+        from importlib import resources
+
+        cand = (Path(str(resources.files("jcm")))
+                / "data" / "bc" / "t63" / "ozone.nc")
+        if cand.exists():
+            ozone_file = str(cand)
+        else:
+            logging.warning(
+                "forcing.ozone_file=auto: packaged t63/ozone.nc missing "
+                "— pySES run falls back to the ANALYTIC ozone profile "
+                "(~12 W/m2 clear-sky OLR low bias)."
+            )
             ozone_file = None
-        if isinstance(ozone_file, str) and "{year}" in ozone_file:
-            # Transient ozone is genuinely unsupported on the pySES path (unlike
-            # oxidants/emissions below): the column ozone leaf is a 12-month
-            # WRAP_YEAR climatology and ``attach_jam_forcing`` rejects any
-            # non-12-month file. Raise the clear limitation here rather than let
-            # the literal-brace path reach ``xr.open_dataset`` as a file-not-
-            # found. Run the spectral dinosaur backend for transient ozone.
-            raise ValueError(
-                f"forcing.ozone_file={ozone_file!r} has a {{year}} pattern, but "
-                "transient ozone is not supported on the pySES backend (the "
-                "column ozone climatology is a 12-month WRAP_YEAR field). "
-                "Provide a single 12-month climatology file, or use the "
-                "spectral dinosaur backend for transient ozone."
-            )
-        provenance.record_fact(
-            "ozone_source",
-            f"prescribed:{ozone_file}" if ozone_file
-            else "analytic (no ozone file)")
-
-        raw_file = _forcing_cfg.get("file", None)
-        if isinstance(raw_file, str) and "{year}" in raw_file:
-            # Transient surface forcing is genuinely unsupported on pySES (as
-            # for ozone above): the column forcing reader opens a SINGLE
-            # 12-month climatology (``jcm.dycore.pyses.forcing.build_forcing``
-            # → one ``xr.open_dataset``), not a multi-year concatenation. Raise
-            # the clear limitation here rather than let the literal-brace path
-            # reach ``_resolve_data_path`` and surface as a confusing hf:// 404
-            # / file-not-found. Use forcing=amip/era5 on the spectral dinosaur
-            # backend for transient surface forcing.
-            raise ValueError(
-                f"forcing.file={raw_file!r} has a {{year}} pattern, but "
-                "transient surface forcing is not supported on the pySES "
-                "backend (the column forcing reader opens a single 12-month "
-                "climatology). Provide a single climatology file, or use the "
-                "spectral dinosaur backend (forcing=amip/era5) for transient "
-                "surface forcing."
-            )
-        file = (_resolve_data_path(raw_file)
-                or _pyses_default_bc("forcing.nc"))
-        # Year expansion must happen on the pySES path too, so the documented
-        # transient forms (``oxidants_file``/``emissions_file=.../{year}.nc``
-        # with ``forcing.years``) resolve to real yearly files rather than a
-        # literal-brace path (or an unfetched ``hf://`` URL) reaching
-        # ``xr.open_dataset``. Oxidants and emissions each go through a shared
-        # resolver that expands ``{year}`` patterns, resolves ``hf://`` and
-        # runs the uniform-time-axis check, so neither can drift from the
-        # spectral path. Both open their file set as ONE combined dataset in
-        # ``attach_jam_forcing`` (pySES has no per-product alignment machinery),
-        # so a genuine multi-product emissions *list* is flattened for that
-        # single open with every ``{year}`` element expanded — and a list that
-        # mixes a climatology with a transient product is rejected there rather
-        # than mis-aligned (the spectral per-product path is the only one that
-        # can carry both in one list). ``dms_file`` / ``dust_file`` take no year
-        # expansion because they are climatology-only on BOTH backends (their
-        # readers are WRAP_YEAR; ``_attach_dms`` / ``_attach_dust`` likewise
-        # never expand); a ``{year}`` there is rejected loudly by
-        # ``_reject_year_pattern`` on both paths rather than reaching
-        # ``open_dataset`` as a literal-brace file-not-found.
-        forcing = pyses_build_forcing(
-            str(file), dycore,
-            emissions_file=_resolve_pyses_emission_paths(_forcing_cfg),
-            dms_file=_resolve_data_path(_reject_year_pattern(
-                _forcing_cfg.get("dms_file", None), "dms_file")),
-            dust_file=_resolve_data_path(_reject_year_pattern(
-                _forcing_cfg.get("dust_file", None), "dust_file")),
-            oxidants_file=_resolve_oxidant_paths(_forcing_cfg),
-            ozone_file=_resolve_data_path(ozone_file),
+    elif ozone_file in ("", "null", "none"):
+        ozone_file = None
+    if isinstance(ozone_file, str) and "{year}" in ozone_file:
+        # Transient ozone is genuinely unsupported on the pySES path (unlike
+        # oxidants/emissions below): the column ozone leaf is a 12-month
+        # WRAP_YEAR climatology and ``attach_jam_forcing`` rejects any
+        # non-12-month file. Raise the clear limitation here rather than let
+        # the literal-brace path reach ``xr.open_dataset`` as a file-not-
+        # found. Run the spectral dinosaur backend for transient ozone.
+        raise ValueError(
+            f"forcing.ozone_file={ozone_file!r} has a {{year}} pattern, but "
+            "transient ozone is not supported on the pySES backend (the "
+            "column ozone climatology is a 12-month WRAP_YEAR field). "
+            "Provide a single 12-month climatology file, or use the "
+            "spectral dinosaur backend for transient ozone."
         )
-        # MACv2-SP plume weights are the one dycore-agnostic attachment the
-        # spectral tail below also performs that ``pyses_build_forcing`` does
-        # NOT: ``aerosol_year_weight``/``aerosol_ann_cycle`` are plume-indexed
-        # scalar time series with NO horizontal field, so they need none of the
-        # column bilinear sampling ``attach_jam_forcing`` does for the gridded
-        # inputs (ozone/emissions/dms/dust/oxidants, which it therefore
-        # reimplements). Reuse the SAME ``_attach_macv2_weights`` helper here so
-        # ``forcing=macv2_sp`` on pySES actually loads its mandatory
-        # ``macv2_file`` instead of silently dropping it — the very silent-ignore
-        # trap warning 4 recommends this config to escape. (The exact-grid
-        # ``validate_emissions_grid``/``validate_oxidant_levels`` checks stay
-        # dinosaur-only: pySES interpolates every field onto columns, so it has
-        # no exact-grid requirement and asserts dim order in ``attach_jam_forcing``
-        # instead. Nudging is likewise dinosaur-only — attached later in ``run``,
-        # not here, and gated off on pySES.)
-        return _attach_macv2_weights(forcing, _forcing_cfg, coords)
+    provenance.record_fact(
+        "ozone_source",
+        f"prescribed:{ozone_file}" if ozone_file
+        else "analytic (no ozone file)")
 
-    forcing_cfg = _forcing_cfg
-    if forcing_cfg is None or forcing_cfg.kind == "default":
-        forcing = None
-    elif forcing_cfg.kind == "from_file":
-        from jcm.forcing import ForcingData
-        files = _expand_years(forcing_cfg.file, forcing_cfg.get("years", None),
-                              forcing_cfg.get("available_years", None))
-        forcing = ForcingData.from_file(
-            _resolve_data_path(files), coords=coords,
-            align_mode=str(forcing_cfg.get("align", "auto")))
-    else:
-        raise ValueError(f"Unknown forcing.kind={forcing_cfg.kind!r}")
-    forcing = _attach_ozone(forcing, forcing_cfg, coords)
-    forcing = _attach_emissions(forcing, forcing_cfg, coords)
-    forcing = _attach_dms(forcing, forcing_cfg, coords)
-    forcing = _attach_dust(forcing, forcing_cfg, coords)
-    forcing = _attach_oxidants(forcing, forcing_cfg, coords)
-    forcing = _attach_macv2_weights(forcing, forcing_cfg, coords)
-    return forcing
-
-
-def _model_latlon_deg(coords):
-    """Model nodal latitudes/longitudes in degrees (dinosaur stores radians)."""
-    import numpy as np
-    lat_deg = np.asarray(coords.horizontal.latitudes) * 180.0 / np.pi
-    lon_deg = np.asarray(coords.horizontal.longitudes) * 180.0 / np.pi
-    return lat_deg, lon_deg
-
-
-def _ensure_parent_forcing(forcing, coords):
-    """Build the aquaplanet parent ``ForcingData`` when ``kind: default``.
-
-    Same rationale as ``_attach_ozone``: ``default_forcing`` preserves the
-    cos²-latitude SST climatology that ``ForcingData.zeros`` would silently
-    replace with a uniform 288.15 K placeholder.
-    """
-    if forcing is not None:
-        return forcing
-    from jcm.forcing import default_forcing
-    return default_forcing(coords.horizontal)
-
-
-def _grid_token(coords) -> str:
-    """Mirror grid token (``"t63"``) for the model's horizontal grid.
-
-    Derived from the spectral resolution (truncation =
-    ``total_wavenumbers - 2``, the same relation ``utils.get_coords``
-    uses), so no hand-maintained table can go stale; whether the mirror
-    actually carries the grid is decided by the fetch itself.
-    """
-    return bundle_names.grid_token(
-        int(coords.horizontal.total_wavenumbers) - 2)
-
-
-def _vertical_kind(coords) -> str:
-    """Coordinate family (``"hybrid"``/``"sigma"``) of the model's vertical grid.
-
-    Gates the level-resolved ``auto`` products (oxidants, ozone), which the data
-    mirror interpolates onto hybrid-level pressures and maps level-for-level: a
-    sigma grid must NOT pull a hybrid bundle that merely shares its (token, nlev).
-    Uses the same ``HybridCoordinates`` test as
-    :func:`jcm.forcing.validate_oxidant_levels` so the availability gate and the
-    (sigma-skipping) coefficient validator classify a grid identically.
-    """
-    from dinosaur.hybrid_coordinates import HybridCoordinates
-    return "hybrid" if isinstance(coords.vertical, HybridCoordinates) else "sigma"
+    raw_file = _forcing_cfg.get("file", None)
+    if isinstance(raw_file, str) and "{year}" in raw_file:
+        # Transient surface forcing is genuinely unsupported on pySES (as
+        # for ozone above): the column forcing reader opens a SINGLE
+        # 12-month climatology (``jcm.dycore.pyses.forcing.build_forcing``
+        # → one ``xr.open_dataset``), not a multi-year concatenation. Raise
+        # the clear limitation here rather than let the literal-brace path
+        # reach ``_resolve_data_path`` and surface as a confusing hf:// 404
+        # / file-not-found. Use forcing=amip/era5 on the spectral dinosaur
+        # backend for transient surface forcing.
+        raise ValueError(
+            f"forcing.file={raw_file!r} has a {{year}} pattern, but "
+            "transient surface forcing is not supported on the pySES "
+            "backend (the column forcing reader opens a single 12-month "
+            "climatology). Provide a single climatology file, or use the "
+            "spectral dinosaur backend (forcing=amip/era5) for transient "
+            "surface forcing."
+        )
+    file = (_resolve_data_path(raw_file)
+            or _pyses_default_bc("forcing.nc"))
+    # Year expansion must happen on the pySES path too, so the documented
+    # transient forms (``oxidants_file``/``emissions_file=.../{year}.nc``
+    # with ``forcing.years``) resolve to real yearly files rather than a
+    # literal-brace path (or an unfetched ``hf://`` URL) reaching
+    # ``xr.open_dataset``. Oxidants and emissions each go through a shared
+    # resolver that expands ``{year}`` patterns, resolves ``hf://`` and
+    # runs the uniform-time-axis check, so neither can drift from the
+    # spectral path. Both open their file set as ONE combined dataset in
+    # ``attach_jam_forcing`` (pySES has no per-product alignment machinery),
+    # so a genuine multi-product emissions *list* is flattened for that
+    # single open with every ``{year}`` element expanded — and a list that
+    # mixes a climatology with a transient product is rejected there rather
+    # than mis-aligned (the spectral per-product path is the only one that
+    # can carry both in one list). ``dms_file`` / ``dust_file`` take no year
+    # expansion because they are climatology-only on BOTH backends (their
+    # readers are WRAP_YEAR; ``_attach_dms`` / ``_attach_dust`` likewise
+    # never expand); a ``{year}`` there is rejected loudly by
+    # ``_reject_year_pattern`` on both paths rather than reaching
+    # ``open_dataset`` as a literal-brace file-not-found.
+    forcing = pyses_build_forcing(
+        str(file), dycore,
+        emissions_file=_resolve_pyses_emission_paths(_forcing_cfg),
+        dms_file=_resolve_data_path(_reject_year_pattern(
+            _forcing_cfg.get("dms_file", None), "dms_file")),
+        dust_file=_resolve_data_path(_reject_year_pattern(
+            _forcing_cfg.get("dust_file", None), "dust_file")),
+        oxidants_file=_resolve_oxidant_paths(_forcing_cfg),
+        ozone_file=_resolve_data_path(ozone_file),
+    )
+    # MACv2-SP plume weights are the one dycore-agnostic attachment the
+    # spectral tail below also performs that ``pyses_build_forcing`` does
+    # NOT: ``aerosol_year_weight``/``aerosol_ann_cycle`` are plume-indexed
+    # scalar time series with NO horizontal field, so they need none of the
+    # column bilinear sampling ``attach_jam_forcing`` does for the gridded
+    # inputs (ozone/emissions/dms/dust/oxidants, which it therefore
+    # reimplements). Reuse the SAME ``_attach_macv2_weights`` helper here so
+    # ``forcing=macv2_sp`` on pySES actually loads its mandatory
+    # ``macv2_file`` instead of silently dropping it — the very silent-ignore
+    # trap warning 4 recommends this config to escape. (The exact-grid
+    # ``validate_emissions_grid``/``validate_oxidant_levels`` checks stay
+    # dinosaur-only: pySES interpolates every field onto columns, so it has
+    # no exact-grid requirement and asserts dim order in ``attach_jam_forcing``
+    # instead. Nudging is likewise dinosaur-only — attached later in ``run``,
+    # not here, and gated off on pySES.)
+    return _attach_macv2_weights(forcing, _forcing_cfg, coords)
 
 
 def _resolve_auto_ozone(coords):
@@ -1320,495 +1132,6 @@ def _resolve_auto_terrain(coords):
             f"({nlon}x{nlat}) and the mirror fetch of "
             f"bundles/{token}/terrain.nc failed: {e}"
         ) from e
-
-
-def _attach_ozone(forcing, forcing_cfg, coords):
-    """Load the ozone climatology and attach to ``forcing``.
-
-    ``ozone_file: auto`` (the shipped default) resolves a packaged
-    climatology matching the grid via ``_resolve_auto_ozone``; no match
-    degrades to the analytic ozone profile with a warning. An explicit
-    path is loaded strictly (errors on any mismatch). ``null`` disables
-    the climatology silently (analytic profile, no warning).
-
-    When ``forcing`` is ``None`` (``kind: default``) and an ozone file IS
-    given, build the parent struct via ``default_forcing(...)`` so the
-    aquaplanet cos²-latitude SST climatology is preserved — using
-    ``ForcingData.zeros`` here would silently swap it for the uniform
-    288.15 K placeholder, materially changing the boundary conditions
-    for any run configured with only ``ozone_file``.
-    """
-    if forcing_cfg is None:
-        return forcing
-    ozone_file = _resolve_data_path(_expand_years(
-        forcing_cfg.get("ozone_file", None),
-        forcing_cfg.get("years", None),
-        _product_available_years(forcing_cfg, "ozone_available_years")))
-    if isinstance(ozone_file, (list, tuple)):
-        ozone_file = [str(p) for p in ozone_file]
-    if ozone_file in (None, "", "null"):
-        provenance.record_fact("ozone_source", "analytic (no ozone_file)")
-        return forcing
-    if ozone_file == "auto":
-        ozone_file = _resolve_auto_ozone(coords)
-        if ozone_file is None:
-            provenance.record_fact(
-                "ozone_source", "analytic (auto found no packaged match)")
-            logging.warning(
-                "forcing.ozone_file=auto: no packaged jcm/data/bc/*/ozone.nc "
-                "matches this grid — falling back to the ANALYTIC ozone "
-                "profile, whose ~7.6x tropospheric ozone column biases "
-                "clear-sky OLR low by ~12 W/m2. Prepare a climatology with "
-                "jcm.data.bc.interpolate_ozone for production radiation."
-            )
-            return forcing
-        logging.info("forcing.ozone_file=auto resolved to %s", ozone_file)
-    import numpy as np
-
-    from jcm.forcing import default_forcing
-    from jcm.ozone_climatology import OzoneClimatology
-    nlon, nlat = coords.horizontal.nodal_shape
-    nlev = coords.nodal_shape[0]
-    # Pass the model's lat/lon (degrees) so the loader catches files
-    # with the right shape but flipped/shifted grids — same N points,
-    # wrong column mapping, would otherwise wire ozone into the wrong
-    # latitudes silently. Dinosaur stores both in radians.
-    lat_deg = np.asarray(coords.horizontal.latitudes) * 180.0 / np.pi
-    lon_deg = np.asarray(coords.horizontal.longitudes) * 180.0 / np.pi
-    climatology = OzoneClimatology.from_file(
-        ozone_file,
-        nlon=int(nlon), nlat=int(nlat), nlev=int(nlev),
-        lat_deg=lat_deg, lon_deg=lon_deg,
-    )
-    provenance.record_fact("ozone_source", f"prescribed:{ozone_file}")
-    provenance.record_input(ozone_file)
-    if forcing is None:
-        forcing = default_forcing(coords.horizontal)
-    return forcing.copy(ozone_climatology=climatology)
-
-
-def _merge_disjoint_emissions(acc, acc_src, new, path):
-    """Merge one emission product's variables into ``acc``, rejecting overlaps.
-
-    Emission products are contractually **disjoint**: each ``emissions_file``
-    list entry contributes distinct ``emis_<sector>_<species>`` /
-    ``aero_emis_<tracer>`` variables (see :func:`_attach_emissions`). Two
-    products claiming the SAME variable is ambiguous user intent, so raise a
-    build-time ``ValueError`` naming the colliding variable(s) and both
-    products — the runner will neither silently keep the last (``dict.update``
-    is last-one-wins, the F1 defect) nor invent an additive merge (summing two
-    products for one sector/species would silently double-count). ``acc_src``
-    maps each already-merged variable to the product that supplied it, purely
-    to name both sides of a collision.
-    """
-    collisions = [v for v in new if v in acc]
-    if collisions:
-        detail = "; ".join(
-            f"{v!r} (from {acc_src[v]!r} and {str(path)!r})"
-            for v in collisions)
-        raise ValueError(
-            "forcing.emissions_file: duplicate emission variable(s) across "
-            f"products — {detail}. Each list entry must contribute DISTINCT "
-            "emission variables (different sectors/species); two products "
-            "claiming the same variable is ambiguous, so the runner neither "
-            "silently keeps one nor sums them (which would double-count). "
-            "Remove the overlap, or merge the overlapping products offline "
-            "into a single file."
-        )
-    acc.update(new)
-    acc_src.update({v: str(path) for v in new})
-
-
-def _attach_emissions(forcing, forcing_cfg, coords):
-    """Attach prescribed aerosol emissions from ``cfg.forcing.emissions_file``.
-
-    No-op when unset. ``emissions_file`` may be a single path or a **list** of
-    paths (e.g. one file for biomass burning and one for the rest). Each list
-    element is a **product** opened and time-aligned on its own (see
-    :func:`_forcing_products`), so a transient ``{year}`` product and a 12-month
-    climatology can be mixed without outer-joining their disjoint time axes; the
-    per-variable ``TimeSeries`` leaves from all products merge into one
-    ``ForcingData``, each keeping its own time axis. The fields auto-route by
-    content: variables named ``emis_<sector>_<species>`` drive the bulk /
-    in-model-speciated path (``anthropogenic_emissions``); ``aero_emis_<tracer>``
-    variables drive the CAM6-faithful pre-speciated path
-    (``prescribed_aerosol_emissions``). A file may carry either or both. The
-    fields must already be on the model horizontal
-    grid — this does **not** regrid (use :mod:`jcm.data.emissions.prepare`
-    first); a grid mismatch raises rather than silently zeroing (the emission
-    terms fall back to zero on a size mismatch, which from the CLI would look
-    like the file "did nothing"). Like ozone, when ``kind: default`` supplies no
-    parent struct one is built via ``default_forcing`` so the aquaplanet SST
-    climatology is preserved.
-
-    The matching emission term must also be in the physics package (e.g.
-    ``physics=echam-jam``) for the fields to be consumed.
-    """
-    if forcing_cfg is None:
-        return forcing
-    raw = forcing_cfg.get("emissions_file", None)
-    if raw in (None, "", "null"):
-        return forcing
-    years = forcing_cfg.get("years", None)
-    # Per-product coverage: a preset may run its SURFACE forcing past the
-    # emissions series' end (``forcing=era5`` runs to 2024 but the mirror's
-    # ``emissions_amip`` bundle ends 2022), so honour an
-    # ``emissions_available_years`` override — the same mechanism ozone uses —
-    # and only fall back to the shared ``available_years`` when it is unset.
-    # Without this, ``emissions_file=.../{year}.nc`` would over-expand into
-    # never-built 2023/2024 files following the transient-warning's advice.
-    available = _product_available_years(
-        forcing_cfg, "emissions_available_years")
-
-    from jcm.forcing import (
-        default_forcing,
-        read_anthropogenic_emissions,
-        read_prescribed_aerosol_emissions,
-        validate_emissions_grid,
-    )
-
-    # Read each product independently and merge the per-variable TimeSeries
-    # leaves. Each product keeps its own time axis / align_mode, so mixing a
-    # transient product with a climatology does not force one shared axis.
-    #
-    # This per-product merge is meaningful HERE but not for oxidants (contrast
-    # _attach_oxidants, which handles a list as ONE product): emission products
-    # carry DISJOINT variables (different ``emis_<sector>_<species>`` /
-    # ``aero_emis_<tracer>`` sets), so merging across products unions genuinely
-    # distinct keys. That disjointness is a contract, not a hope:
-    # ``_merge_disjoint_emissions`` REJECTS a variable supplied by two products
-    # (F1) rather than let a plain ``dict.update`` silently keep the last.
-    # Oxidant files must each carry the IDENTICAL four gases, so an analogous
-    # update would be pure last-one-wins.
-    anthro: dict = {}
-    speciated: dict = {}
-    # Provenance for the disjoint-merge check: which product supplied each
-    # already-merged variable, for a precise collision message (F1).
-    anthro_src: dict = {}
-    speciated_src: dict = {}
-    for product in _forcing_products(raw, years, available):
-        path = _resolve_data_path(product)
-        if path in (None, "", "null"):
-            continue
-        ds = _open_forcing_dataset(path)
-        try:
-            a = read_anthropogenic_emissions(ds)
-            s = read_prescribed_aerosol_emissions(ds)
-        finally:
-            ds.close()
-        if a is None and s is None:
-            raise ValueError(
-                f"forcing.emissions_file {path!r} has no emissions variables: "
-                "expected ``emis_<sector>_<species>`` (bulk) or "
-                "``aero_emis_<tracer>`` (pre-speciated). See the emissions-file "
-                "contract in docs/design/jam.md."
-            )
-        if a:
-            _merge_disjoint_emissions(anthro, anthro_src, a, path)
-        if s:
-            _merge_disjoint_emissions(speciated, speciated_src, s, path)
-    if not anthro and not speciated:
-        return forcing
-    validate_emissions_grid({**anthro, **speciated}, coords, raw)
-    if forcing is None:
-        forcing = default_forcing(coords.horizontal)
-    return forcing.copy(anthropogenic_emissions=anthro or None,
-                        prescribed_aerosol_emissions=speciated or None)
-
-
-def _reject_year_pattern(value, key):
-    """Fail loudly on a ``{year}`` pattern for a climatology-only forcing key.
-
-    ``dms_file`` and ``dust_file`` are single-file WRAP_YEAR climatologies: the
-    data mirror publishes NO transient (yearly) DMS or dust product, and their
-    readers (:func:`jcm.forcing.read_dms_seawater` /
-    :func:`jcm.forcing.read_dust_source`) do not expand ``{year}``. A pattern
-    here would otherwise reach ``xr.open_dataset`` as a literal-brace path and
-    die with a cryptic file-not-found; reject it up front, naming the reason and
-    that only ``emissions_file`` / ``oxidants_file`` accept yearly patterns.
-    Returns ``value`` unchanged when it is not a pattern.
-    """
-    if isinstance(value, str) and "{year}" in value:
-        raise ValueError(
-            f"forcing.{key}={value!r} contains a {{year}} pattern, but {key} "
-            "is a climatology-only single file: the data mirror publishes no "
-            "transient (yearly) DMS/dust product and its reader does not expand "
-            "{year}. Only forcing.emissions_file and forcing.oxidants_file "
-            "accept yearly patterns; give dms_file/dust_file a single 12-month "
-            "climatology (or 'auto'/'null').")
-    return value
-
-
-def _attach_dms(forcing, forcing_cfg, coords):
-    """Attach the seawater-DMS climatology from ``cfg.forcing.dms_file``.
-
-    No-op when unset. Loads a HAMMOZ-style monthly ``DMS_sea (time, lat, lon)``
-    climatology (nmol/L, converted to kg/m³ — see
-    :func:`jcm.forcing.read_dms_seawater`) as a ``WRAP_YEAR`` ``TimeSeries``
-    on ``forcing.dms_seawater``, which :class:`DmsEmissions` consumes. The
-    file must already be on the model horizontal grid; lat/lon values are
-    validated (a descending-latitude file is flipped) and a mismatch raises —
-    the term otherwise falls back to zero on a size mismatch, which from the
-    CLI would look like the file "did nothing". Needs a JAM physics package
-    (e.g. ``physics=echam-jam``) for the field to be consumed.
-    """
-    if forcing_cfg is None:
-        return forcing
-    path = _resolve_data_path(
-        _reject_year_pattern(forcing_cfg.get("dms_file", None), "dms_file"))
-    if path in (None, "", "null"):
-        return forcing
-    import xarray as xr
-
-    from jcm.forcing import read_dms_seawater
-    lat_deg, lon_deg = _model_latlon_deg(coords)
-    with xr.open_dataset(str(path)) as ds:
-        ts = read_dms_seawater(ds, lat_deg=lat_deg, lon_deg=lon_deg)
-    forcing = _ensure_parent_forcing(forcing, coords)
-    return forcing.copy(dms_seawater=ts)
-
-
-def _attach_dust(forcing, forcing_cfg, coords):
-    """Attach the dust-source/erodibility map from ``cfg.forcing.dust_file``.
-
-    No-op when unset. Loads a HAMMOZ-style monthly ``pot_source
-    (time, lat, lon)`` climatology (clipped to the [0, 1] erodibility contract
-    of :class:`DustEmissions` — see :func:`jcm.forcing.read_dust_source`) as a
-    ``WRAP_YEAR`` ``TimeSeries`` on ``forcing.dust_source``. Grid handling as
-    in :func:`_attach_dms`.
-    """
-    if forcing_cfg is None:
-        return forcing
-    path = _resolve_data_path(
-        _reject_year_pattern(forcing_cfg.get("dust_file", None), "dust_file"))
-    if path in (None, "", "null"):
-        return forcing
-    import xarray as xr
-
-    from jcm.forcing import read_dust_source
-    lat_deg, lon_deg = _model_latlon_deg(coords)
-    with xr.open_dataset(str(path)) as ds:
-        ts = read_dust_source(ds, lat_deg=lat_deg, lon_deg=lon_deg)
-    forcing = _ensure_parent_forcing(forcing, coords)
-    return forcing.copy(dust_source=ts)
-
-
-def _resolve_oxidant_paths(forcing_cfg):
-    """Resolve ``forcing.oxidants_file`` into the single oxidant product's files.
-
-    Shared by the spectral (:func:`_attach_oxidants`) and pySES
-    (:func:`_build_forcing`) paths so ``{year}`` expansion, ``hf://`` resolution
-    and the uniform-time-axis validation cannot drift between the two — the
-    forked bypass that previously let the pySES branch hand a literal
-    ``{year}`` brace path (or an unfetched ``hf://`` URL) straight to
-    ``xr.open_dataset``. A ``{year}`` pattern expands to the product's yearly
-    files; an explicit list is taken verbatim as that one product's file set;
-    either way the set is opened together (``open_mfdataset``, by-coords)
-    downstream, so :func:`_assert_uniform_time_axis` runs here to reject
-    a mixed set up front. Returns a non-empty list of string paths, or ``None``
-    when ``oxidants_file`` is unset/empty.
-    """
-    if forcing_cfg is None:
-        return None
-    raw = forcing_cfg.get("oxidants_file", None)
-    if raw in (None, "", "null"):
-        return None
-    from omegaconf import ListConfig
-    # Per-product coverage, resolved HERE so both the spectral and pySES paths
-    # (which share this helper) clamp a transient oxidant ``{year}`` pattern to
-    # the same range and cannot drift. The mirror publishes no transient
-    # oxidants product, but a user bringing their own series whose coverage
-    # differs from the surface forcing's sets ``oxidants_available_years``; it
-    # falls back to the shared ``available_years`` when unset.
-    files = _resolve_data_path(_expand_years(
-        raw, forcing_cfg.get("years", None),
-        _product_available_years(forcing_cfg, "oxidants_available_years")))
-    if isinstance(files, (list, tuple, ListConfig)):
-        paths = [str(p) for p in files
-                 if str(p) not in ("", "null", "none", "None")]
-    elif files in (None, "", "null"):
-        return None
-    else:
-        paths = [str(files)]
-    if not paths:
-        return None
-    # Oxidants are ONE product (its ``{year}`` expansion or explicit file set),
-    # so the whole list is a single product for the guard: a lone climatology
-    # or one transient product's yearly files are uniform, while a file set
-    # straddling integer-month and datetime axes is rejected as malformed.
-    _assert_uniform_time_axis([paths], config_key="forcing.oxidants_file")
-    return paths
-
-
-def _resolve_pyses_emission_paths(forcing_cfg):
-    """Resolve ``forcing.emissions_file`` into a flat file list for pySES.
-
-    The pySES column backend opens ALL emission files as ONE combined dataset
-    (``jcm.dycore.pyses.forcing.attach_jam_forcing`` → a single
-    ``open_mfdataset``, by-coords), so every element — a scalar path, an
-    ``hf://`` URL, or a ``{year}`` pattern — is expanded and the result
-    FLATTENED into one path list for that single open. A ``{year}`` element
-    becomes its product's yearly files (via :func:`_forcing_products`, the same
-    expansion the spectral path uses), so both a scalar ``{year}`` pattern and a
-    ``{year}`` element *inside a list* resolve to real yearly files rather than a
-    literal-brace path reaching ``xr.open_dataset``.
-
-    This deliberately differs from the spectral :func:`_attach_emissions`, which
-    opens each list element as an INDEPENDENT product with its own time axis
-    (per-product ``TimeSeries`` alignment) and can therefore mix a transient
-    ``{year}`` product with a 12-month climatology in one list. pySES has no
-    per-product alignment machinery — one combined open along a single time
-    axis — so a mixed climatology+transient list cannot be aligned and is
-    rejected up front by :func:`_assert_uniform_time_axis` (the honest
-    semantics here), exactly as the shared oxidant path does. Returns a
-    non-empty list of string paths, or ``None`` when ``emissions_file`` is
-    unset/empty.
-    """
-    if forcing_cfg is None:
-        return None
-    raw = forcing_cfg.get("emissions_file", None)
-    if raw in (None, "", "null"):
-        return None
-    years = forcing_cfg.get("years", None)
-    available = _product_available_years(
-        forcing_cfg, "emissions_available_years")
-    # Keep each list element / ``{year}`` expansion as its own product for the
-    # uniform-time-axis guard (the time axis is a per-product property — see
-    # :func:`_product_time_axis`), then flatten to the single path list
-    # ``attach_jam_forcing`` opens by-coords.
-    products: list[list[str]] = []
-    for product in _forcing_products(raw, years, available):
-        resolved = _resolve_data_path(product)
-        if isinstance(resolved, (list, tuple)):
-            files = [str(p) for p in resolved
-                     if str(p) not in ("", "null", "none", "None")]
-        elif resolved not in (None, "", "null"):
-            files = [str(resolved)]
-        else:
-            files = []
-        if files:
-            products.append(files)
-    if not products:
-        return None
-    _assert_uniform_time_axis(products, config_key="forcing.emissions_file")
-    return [p for product in products for p in product]
-
-
-def _attach_oxidants(forcing, forcing_cfg, coords):
-    """Attach the oxidant climatology from ``cfg.forcing.oxidants_file``.
-
-    No-op when unset. Loads a HAMMOZ/MACC-style monthly
-    ``OH/NO3/O3/H2O2_VMR_avrg (time, mlev, lat, lon)`` mole-fraction
-    climatology on ECHAM hybrid model levels into ``forcing.oxidant_vmr`` as
-    ``WRAP_YEAR`` ``TimeSeries`` leaves; :class:`PrescribedOxidants` converts
-    VMR → molec cm⁻³ in-term where T and p are available. The file's levels
-    are mapped **one-to-one** onto the model levels: the level count is
-    asserted in :func:`jcm.forcing.read_oxidant_vmr`, and when the model runs
-    hybrid vertical coordinates the file's ``hyam``/``hybm`` are additionally
-    cross-checked against the model's coefficients here, so a file on
-    different 47 levels can't be wired in silently. Horizontal grid handling
-    as in :func:`_attach_dms`.
-
-    Supports the same yearly-expansion the anthropogenic emissions path does, so
-    the year-matched transient oxidant product recommended for a transient run
-    (``oxidants_file=.../{year}.nc`` with ``forcing.years``) actually loads: a
-    ``{year}`` pattern expands to one file per year, which are concatenated
-    along the time axis (``open_mfdataset``, by-coords) and read with ``auto``
-    alignment — a single 12-month climatology stays ``WRAP_YEAR`` while a
-    multi-year axis becomes ``BY_DATE``. The level-for-level vertical mapping
-    is unchanged (the yearly files share the model's hybrid grid).
-
-    Oxidants are handled as **one product**, unlike the per-product emissions
-    path. A user-supplied **list** ``oxidants_file`` means the yearly files of a
-    *single* oxidant product (exactly what ``{year}`` expansion produces): the
-    whole file set is opened together (``open_mfdataset``, by-coords) along one
-    time axis and read once. This differs deliberately from
-    :func:`_attach_emissions`, whose per-product merge is meaningful because
-    emission products carry **disjoint** variables (different sectors/species).
-    :func:`jcm.forcing.read_oxidant_vmr` instead requires *every* oxidant file
-    to carry **all four** gases (oh/no3/o3/h2o2), so distinct products fully
-    overlap — a per-product ``dict.update`` would be pure last-one-wins and
-    silently keep only the final file. Genuinely incompatible members in one
-    file set (e.g. an integer-month climatology mixed with datetime transients)
-    are rejected up front by :func:`_assert_uniform_time_axis` rather
-    than left to ``open_mfdataset`` to NaN-fill or clash cryptically.
-    """
-    if forcing_cfg is None:
-        return forcing
-    paths = _resolve_oxidant_paths(forcing_cfg)
-    if not paths:
-        return forcing
-
-    import xarray as xr
-
-    from jcm.forcing import read_oxidant_vmr, validate_oxidant_levels
-    lat_deg, lon_deg = _model_latlon_deg(coords)
-    nlev = int(coords.nodal_shape[0])
-
-    # ``data_vars="minimal"`` so only the time-dependent VMR fields are
-    # concatenated across a multi-year set: the static ``hyam``/``hybm`` hybrid
-    # coefficients carry no time axis and must stay 1-D (the default
-    # ``data_vars="all"`` would stack them to a spurious ``(nfiles, mlev)`` that
-    # breaks ``read_oxidant_vmr`` / ``validate_oxidant_levels``).
-    ds = (xr.open_mfdataset(paths, combine="by_coords", data_vars="minimal")
-          if len(paths) > 1 else xr.open_dataset(paths[0]))
-    ref = paths if len(paths) > 1 else paths[0]
-    try:
-        mapping = read_oxidant_vmr(ds, nlev=nlev, lat_deg=lat_deg,
-                                   lon_deg=lon_deg, align_mode="auto")
-        validate_oxidant_levels(ds, coords, ref)
-    finally:
-        ds.close()
-    if not mapping:
-        return forcing
-    forcing = _ensure_parent_forcing(forcing, coords)
-    return forcing.copy(oxidant_vmr=mapping)
-
-
-def _product_time_axis(product):
-    """One product's combined time axis (thin adapter over the engine).
-
-    See :func:`jcm.data.input_resolution.product_time_axis` — the single home
-    for the merge-key computation shared with the Python door.
-    """
-    return ir.product_time_axis(product)
-
-
-def _assert_uniform_time_axis(products, *, config_key) -> None:
-    """Reject a product set that cannot share one by-coords time axis.
-
-    Thin adapter over :func:`jcm.data.input_resolution.assert_uniform_time_axis`
-    — the merge-compatibility rule lives once in the engine (#750/#751).
-    """
-    ir.assert_uniform_time_axis(products, config_key=config_key)
-
-
-def _attach_macv2_weights(forcing, forcing_cfg, coords):
-    """Attach time-varying MACv2-SP plume weights from ``forcing.macv2_file``.
-
-    No-op when unset. Loads the ``year_weight``/``ann_cycle`` scalings from a
-    ``MACv2.0-SP_v1.nc`` file (via :func:`jcm.forcing.read_macv2_weights`) onto
-    ``forcing.aerosol_year_weight`` / ``aerosol_ann_cycle`` — the fields the
-    MACv2-SP aerosol term reads for per-year amplitude and the seasonal cycle.
-    Without a file these default to all-ones (perpetual year-2005 amplitude, no
-    seasonal cycle); ``forcing=macv2_sp`` sets the key. ``macv2_file=auto`` (that
-    config's default) resolves to the repo-packaged SPv2.1 file
-    (:func:`jcm.forcing.packaged_macv2_path`); an explicit path overrides. The
-    weights are plume-indexed and grid-independent, so no regridding is needed.
-    """
-    if forcing_cfg is None:
-        return forcing
-    raw = forcing_cfg.get("macv2_file", None)
-    if raw in (None, "", "null"):
-        return forcing
-    from jcm.forcing import packaged_macv2_path
-    path = _resolve_data_path(
-        packaged_macv2_path() if raw == "auto" else raw)
-    from jcm.forcing import read_macv2_weights
-    year_weight, ann_cycle = read_macv2_weights(str(path))
-    forcing = _ensure_parent_forcing(forcing, coords)
-    return forcing.copy(aerosol_year_weight=year_weight,
-                        aerosol_ann_cycle=ann_cycle)
 
 
 # ---------------------------------------------------------------------------
